requirements.txt
# ══════════════════════════════════════════
# AgenteInteligência-AI — Dependências
# Instalar: pip install -r requirements.txt
# ══════════════════════════════════════════

# === Core ===
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
jinja2>=3.1.0
aiofiles>=23.0.0

# === Database ===
aiosqlite>=0.20.0

# === LangChain / LangGraph ===
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# === LLM Providers ===
openai>=1.40.0
httpx>=0.27.0

# === Observabilidade ===
langfuse>=2.0.0

# === Utilities ===
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0.0
websockets>=12.0
